A worrying number of the functions changed are unused now !!

+++ include/graphics.h
@@ -1982,7 +1982,7 @@ extern const u16 gBattleInterface_BallStatusBarPal[];
 extern const u16 gBattleInterface_BallDisplayPal[];
 
 extern const u32 gSwapLineGfx[];
-extern const u32 gSwapLinePal[];
+extern const u16 gSwapLinePal[];
 extern const u32 gSwapLineGfx_RegisteredItemsMenu[];
 
 extern const u32 gBattlePyramidBag_Gfx[];
